rename: pages 폴더 오타 수정
