+++ front/src/App.tsx
@@ -1,10 +1,10 @@
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import AuthProvider from './context/auth-context';
 import NoneLayout from './components/Layouts/NoneLayout';
-import MyProfile from './components/MyProfile/MyProfile';
 import Intro from './components/Intro/Intro';
 import BaseLayout from './components/Layouts/BaseLayout';
-import MainPage from './page/MainPage';
+import MainPage from './pages/MainPage';
+import MyProfilePage from './pages/MyProfilePage';
 
 function App() {
   return (
@@ -22,7 +22,7 @@ function App() {
               <Route path="/letters" element={<MainPage />} />
               <Route path="/followings" element={<MainPage />} />
               <Route path="/voca" element={<MainPage />} />
-              <Route path="/my-profile" element={<MyProfile />} />
+              <Route path="/my-profile" element={<MyProfilePage />} />
             </Route>
           </Routes>
         </AuthProvider>
