+++ front/src/pages/MyProfilePage.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const MyProfilePage = () => {
+  return <div>MyProfilePage</div>;
+};
+
+export default MyProfilePage;
